DiccionarioDjango/requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
dj-database-url==2.3.0
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.1
gunicorn==23.0.0
joblib==1.4.2
nltk==3.8.1
packaging==24.2
psycopg2==2.9.10
pyspellchecker==0.8.1
regex==2024.5.15
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
whitenoise==6.8.2
